# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Microservice to manage PagoPA Backoffice",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "SelfCare Backoffice",
    "version" : "1.4.5"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  }, {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/backoffice/v1"
      },
      "host" : {
        "default" : "api.dev.platform.pagopa.it",
        "enum" : [ "api.dev.platform.pagopa.it", "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
    "/brokers" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y | Read | N\n",
        "operationId" : "getBrokersEC",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "code",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "order by name or code, default = CODE",
          "in" : "query",
          "name" : "orderby",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "CODE"
          }
        }, {
          "in" : "query",
          "name" : "ordering",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "DESC"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokersResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get paginated list of creditor brokers",
        "tags" : [ "Creditor institution's Brokers" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y | Write | N\n",
        "operationId" : "createBroker_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokerResource"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Create a Broker",
        "tags" : [ "Creditor institution's Brokers" ]
      }
    },
    "/brokers/{broker-code}" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "put" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y | Write | N\n",
        "operationId" : "updateBroker",
        "parameters" : [ {
          "description" : "Broker code",
          "in" : "path",
          "name" : "broker-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerEcDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokerDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Update an existing EC broker",
        "tags" : [ "Creditor institution's Brokers" ]
      }
    },
    "/brokers/{broker-code}/stations" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y | Read | N\n",
        "operationId" : "getStationsDetailsListByBroker",
        "parameters" : [ {
          "in" : "path",
          "name" : "broker-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "stationId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StationDetailsResourceList"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get paginated list of stations given a broker code",
        "tags" : [ "Creditor institution's Brokers" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/brokers/{broker-id}/ibans/export" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y | Read | Y\nThe CSV file contains the following columns: `denominazioneEnte, codiceFiscale, iban, stato, dataAttivazioneIban, descrizione, etichetta`",
        "operationId" : "exportIbansToCsv",
        "parameters" : [ {
          "description" : "SelfCare Broker Id. it's an UUID",
          "in" : "path",
          "name" : "broker-id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Export all IBANs of all creditor institutions handled by a broker EC to CSV",
        "tags" : [ "Creditor institution's Brokers" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/psp-brokers" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getBrokersPsp",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Broker's code",
          "in" : "query",
          "name" : "broker-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Broker's name",
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Order by column name",
          "in" : "query",
          "name" : "orderby",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "CODE"
          }
        }, {
          "description" : "Method of sorting",
          "in" : "query",
          "name" : "sorting",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "DESC"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokersPspResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get the list of all PSP brokers",
        "tags" : [ "Payment Service Provider's Brokers" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "createBroker",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerPspDetailsDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokerPspDetailsResource"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Create a new PSP broker",
        "tags" : [ "Payment Service Provider's Brokers" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BrokerDetailsResource" : {
        "required" : [ "extended_fault_bean" ],
        "type" : "object",
        "properties" : {
          "broker_code" : {
            "type" : "string"
          },
          "broker_details" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extended_fault_bean" : {
            "type" : "boolean"
          }
        }
      },
      "BrokerDto" : {
        "type" : "object",
        "properties" : {
          "broker_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "BrokerEcDto" : {
        "required" : [ "brokerCode" ],
        "type" : "object",
        "properties" : {
          "brokerCode" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extendedFaultBean" : {
            "type" : "boolean"
          }
        }
      },
...
      "BrokerPspDetailsDto" : {
        "required" : [ "broker_psp_code", "description", "enabled", "extended_fault_bean" ],
        "type" : "object",
        "properties" : {
          "broker_psp_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extended_fault_bean" : {
            "type" : "boolean"
          }
        }
      },
      "BrokerPspDetailsResource" : {
        "required" : [ "broker_psp_code", "description", "enabled", "extended_fault_bean" ],
        "type" : "object",
        "properties" : {
          "broker_psp_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extended_fault_bean" : {
            "type" : "boolean"
          }
        }
      },
      "BrokerPspResource" : {
        "required" : [ "broker_psp_code", "description", "enabled" ],
        "type" : "object",
        "properties" : {
          "broker_psp_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "BrokerResource" : {
        "type" : "object",
        "properties" : {
          "broker_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extended_fault_bean" : {
            "type" : "boolean"
          }
        }
      },
      "BrokersPspResource" : {
        "type" : "object",
        "properties" : {
          "brokers_psp" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BrokerPspResource"
            }
          },
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
      "BrokersResource" : {
        "type" : "object",
        "properties" : {
          "brokers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BrokerResource"
            }
          },
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
...
      "PageInfo" : {
        "type" : "object",
        "properties" : {
          "items_found" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total_items" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total_pages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "StationDetailResource" : {
        "required" : [ "associatedCreditorInstitutions", "enabled", "primitiveVersion", "stationCode", "version" ],
        "type" : "object",
        "properties" : {
          "activationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "associatedCreditorInstitutions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "brokerCode" : {
            "type" : "string"
          },
          "brokerDescription" : {
            "type" : "string"
          },
          "brokerDetails" : {
            "$ref" : "#/components/schemas/BrokerDetailsResource"
          },
          "brokerObjId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "flagOnline" : {
            "type" : "boolean"
          },
          "ip" : {
            "type" : "string"
          },
          "ip4Mod" : {
            "type" : "string"
          },
          "modifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "pofService" : {
            "type" : "string"
          },
          "port" : {
            "type" : "integer",
            "format" : "int64"
          },
          "port4Mod" : {
            "type" : "integer",
            "format" : "int64"
          },
          "primitiveVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "protocol" : {
            "type" : "string",
            "enum" : [ "HTTPS", "HTTP" ]
          },
          "protocol4Mod" : {
            "type" : "string",
            "enum" : [ "HTTPS", "HTTP" ]
          },
          "proxyEnabled" : {
            "type" : "boolean"
          },
          "proxyHost" : {
            "type" : "string"
          },
          "proxyPassword" : {
            "type" : "string"
          },
          "proxyPort" : {
            "type" : "integer",
            "format" : "int64"
          },
          "proxyUsername" : {
            "type" : "string"
          },
          "redirectIp" : {
            "type" : "string"
          },
          "redirectPath" : {
            "type" : "string"
          },
          "redirectPort" : {
            "type" : "integer",
            "format" : "int64"
          },
          "redirectProtocol" : {
            "type" : "string",
            "enum" : [ "HTTPS", "HTTP" ]
          },
          "redirectQueryString" : {
            "type" : "string"
          },
          "rtInstantaneousDispatch" : {
            "type" : "boolean"
          },
          "service" : {
            "type" : "string"
          },
          "service4Mod" : {
            "type" : "string"
          },
          "stationCode" : {
            "type" : "string"
          },
          "targetHost" : {
            "type" : "string"
          },
          "targetHostPof" : {
            "type" : "string"
          },
          "targetPath" : {
            "type" : "string"
          },
          "targetPathPof" : {
            "type" : "string"
          },
          "targetPort" : {
            "type" : "integer",
            "format" : "int64"
          },
          "targetPortPof" : {
            "type" : "integer",
            "format" : "int64"
          },
          "threadNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timeoutA" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timeoutB" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timeoutC" : {
            "type" : "integer",
            "format" : "int64"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "wrapperStatus" : {
            "type" : "string",
            "enum" : [ "APPROVED", "TO_CHECK", "TO_FIX", "TO_CHECK_UPDATE", "TO_FIX_UPDATE" ]
          }
        }
      },
...
      "StationDetailsResourceList" : {
        "type" : "object",
        "properties" : {
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          },
          "stations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StationDetailResource"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "JWT" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token get after Login",
        "scheme" : "bearer",
        "type" : "http"
      },
      "SubKey" : {
        "description" : "The Azure Subscription Key to access this API.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}